+++ csharp/MapleStory.Test/UserAPI.cs
@@ -10,6 +10,13 @@ namespace MapleStory.Test
         private static readonly MapleStoryAPI api = new MapleStoryAPI(apiKey);
 
         [Test]
+        public async Task GetCharacterList()
+        {
+            var response = await api.GetCharacterList();
+            Console.WriteLine(response.ToJson());
+        }
+
+        [Test]
         public async Task GetAchievement()
         {
             var response = await api.GetAchievement();
