+++ java/src/test/java/UserApi.java
@@ -1,5 +1,5 @@
 import dev.spiralmoon.maplestory.api.kms.MapleStoryApi;
-import dev.spiralmoon.maplestory.api.kms.dto.user.AchievementDTO;
+import dev.spiralmoon.maplestory.api.kms.dto.user.*;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -10,6 +10,13 @@ public class UserApi {
 
     private static final String apiKey = System.getProperty("API_KEY"); // Your API Key
     private static final MapleStoryApi api = new MapleStoryApi(apiKey);
+
+    @Test
+    @DisplayName("캐릭터 목록 조회")
+    void getCharacterList() throws IOException {
+        CharacterListDTO response = api.getCharacterList().join();
+        System.out.println(response.toString());
+    }
 
     @Test
     @DisplayName("업적 정보 조회")
